HDF5: Remove unnecessary casting (#4036)
## Summary
Removes an unnecessary cast and now correctly compares variables of the
same signedness, fixing some compiler warnings.

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate

+++ Src/Extern/HDF5/AMReX_PlotFileUtilHDF5.cpp
@@ -723,7 +723,7 @@ void WriteMultiLevelPlotfileHDF5SingleDset (const std::string& plotfilename,
         if (H5Pget_layout(lev_dcpl_id) == H5D_CHUNKED) {
             hsize_t chunk_size;
             if (H5Pget_chunk(lev_dcpl_id, 1, &chunk_size) > -1) {
-                if ((int)chunk_size > hs_allprocsize[0]) {
+                if (chunk_size > hs_allprocsize[0]) {
                     H5Pset_chunk(lev_dcpl_id, 1, hs_allprocsize);
                 }
             }
@@ -1195,7 +1195,7 @@ void WriteMultiLevelPlotfileHDF5MultiDset (const std::string& plotfilename,
             if (H5Pget_layout(lev_dcpl_id) == H5D_CHUNKED) {
                 hsize_t chunk_size;
                 if (H5Pget_chunk(lev_dcpl_id, 1, &chunk_size) > -1) {
-                    if ((int)chunk_size > hs_allprocsize[0]) {
+                    if (chunk_size > hs_allprocsize[0]) {
                         H5Pset_chunk(lev_dcpl_id, 1, hs_allprocsize);
                     }
                 }
